requirements.txt
pandas
requests
certifi
newsapi-python
langchain
langchain-community
transformers
sentence-transformers
chromadb
huggingface-hub
python-dotenv
